Game Class, Start Button, In-Game background, Settings Buttons
Added basic functionality for the start button temporarily simulated by clicking anywhere on the left half.

Added basic functionality for the settings menu buttons (open and close) temporarily simulated by clicking on the left or right half - of the right half - of the screen.

Once the main menu is designed these buttons will be adjusted to work only when the button image is clicked.

Added background for when the game has started.

Designed "game" class to handle all in game drawing with the run() method (called by draw() in sketch), and keep track of the running state of the game (running or not running).